removed XmlTreeBuilderActions/XmlTreeBuilderStep and moved impls into XmlTreeBuilder
